Move SlidePanel to dedicated directory and update import
